Generate a LabelProvider per labelled type (#1358)
* Remove dead rule-source map from LabelProviderGenerator

attributeToRuleMap was threaded through buildLabelGraph but never read;
the legacy rule `as` syntax it supported was already deleted. Also drops
the orphaned BodyCorpusLabelProvider.java expectation fixture, a fossil
of that era with no remaining references.

Part of TASK-9548 step 1.

* Generate a LabelProvider per labelled type (TASK-9548 step 2)

Widen LabelProviderGenerator to also emit a type-rooted provider for
every type carrying a direct [label ...] on its own, inherited or
overridden attributes, named <ns>.labels.types.<Type>LabelProvider so
it can't collide with the existing function/report-rooted providers
in <ns>.labels. This is what lets a CSV input type's own labels
resolve, since a projection/ingest's function-rooted provider is
rooted at the function's output type instead.

Function and report providers are untouched and stay byte-identical;
they simply keep existing alongside their type-rooted counterparts.

* Stamp @RuneLabelProvider on labelled pojos (TASK-9548 step 3)

ModelObjectGenerator now computes the type-rooted label provider class
via the same LabelProviderGeneratorUtil gate the LabelProviderGenerator
uses, so the annotation and the generated provider class can never
disagree.

* Deprecate function/report label providers in favour of type-rooted ones (TASK-9548 step 4)

Function- and report-rooted providers are still generated in full and remain
load-bearing where a transform's output type is defined upstream, but authors
should prefer the type-rooted provider where one exists.

* Document type-rooted label providers (TASK-9548 step 5)

Add the per-type @RuneLabelProvider case to rune-and-java.mdx, note the
function/report provider's deprecation as a steer (not a removal notice),
and cross-link from reporting.mdx that labels also drive labelled
serialisation (e.g. CSV) for any type, not just report types.

* Address review of type-rooted label providers (TASK-9548)

Review follow-ups on the label-provider branch; no change to which
providers are generated or what labels they resolve.

- Move LabelProviderGenerator and LabelProviderGeneratorUtil to a new
  generator.java.labels package. The gate is now consumed by
  ModelObjectGenerator too, so a "function" package no longer describes
  it, and the object -> function package dependency goes away.
- Emit {@code @RuneLabelProvider} as plain javadoc text rather than an
  {@link}, so generated function/report providers no longer carry an
  import that only javadoc uses. Their diff against main is now exactly
  the deprecation header, as intended.
- Drop the unused RAttribute import left behind when the dead
  rule-source map was removed.
- Pin the emitted source of a type-rooted provider with an expectation
  file (type-flat/FooLabelProvider.java): package, the absence of the
  @Deprecated header, and the whitespace the Xtend template is
  sensitive to. This is the parity gate for migrating the generator.
- Add TypeLabelProviderDiscoveryTest: generates, compiles and loads a
  flat labelled type with no transform or report anywhere in the model
  (the CSV shape), then asserts the pojo's @RuneLabelProvider value is
  the provider class actually generated. The annotation and the class
  come from two generators off one gate, so nothing else fails if they
  drift - it would surface only as a missing symbol downstream.
- Correct the claim that a choice's options cannot carry labels: a
  ChoiceOption can carry [label ...] (Rosetta.xtext:80-82) and a
  labelled choice does get a provider. Pinned by a new test; the
  misleading comment in testComplexReportLabels is fixed.
- Rename assertSingleGeneratedFile to assertGeneratedFileMatches, which
  is what it does now that it selects by path.
- Use a relative docs cross-link, matching the rest of the website.

* Give the function/report deprecation an expiry (TASK-9548)

Reframe the @Deprecated on function/report label providers as
scheduled for removal at the next major version, instead of implying
they stay forever. The fallback is still load-bearing until then
(an upstream artifact may not have a type-rooted provider to fall
back to), but "no removal ever planned" is what made the annotation
noise rather than a real deprecation.

* Revert function/report provider deprecation (TASK-9548)

Simon's second review pass on #1358 showed the "next major version"
expiry had no real mechanism: the type-provider gate is deliberately
"direct labels only", so an output type whose labels are all on
nested descendants never gets a type-rooted provider in any DSL
version - regenerating dependencies doesn't close that gap, because
it's structural, not staleness.

Rather than widen the gate or add a resolver-side fallback, drop the
deprecation: the function/report provider is permanently necessary
for transform functions and reports, independent of the type-rooted
provider, which exists mainly for serialiser consumers working
outside a transform (e.g. CSV against an already-flat type). The two
coexist permanently and neither supersedes the other.

* Remove internal plan section references from test comment

Comment cited plan-document section numbers (§2.5/§3.2) that mean
nothing outside the plan file; reworded to stand on its own.
